requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.6
python-dotenv>=1.0
sqlalchemy>=2.0
asyncpg>=0.29
alembic>=1.13
httpx>=0.27
requests>=2.32
pyyaml>=6.0
